+++ app/services/catalog/product_service.py
@@ -1,4 +1,5 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException, status
 from app.db.repositories.catalog import product_repository as repo
 from app.schemas.catalog.product import (
     ProductCreate,
@@ -6,7 +7,7 @@ from app.schemas.catalog.product import (
     ProductDelete,
     ProductView,
 )
-from app.schemas.common.query_params import BaseQueryParams
+from app.schemas.common import BaseQueryParams
 
 
 def get_all_products(db: Session, query: BaseQueryParams | None = None):
@@ -20,10 +21,20 @@ def get_all_products(db: Session, query: BaseQueryParams | None = None):
 
 
 def get_product_by_id(db: Session, id: int):
+    if not repo.exist_product_id(db=db, id=id):
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Product not found",
+        )
+
     result = repo.get_by_id_with_category_name(db=db, id=id)
 
     if not result:
-        return None
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Product not found",
+        )
+        
     product, category_name = result
 
     return ProductView(
@@ -44,11 +55,25 @@ def create_product(product_in: ProductCreate, db: Session, user_id: int):
 
 def update_product(req: ProductUpdate, db: Session, user_id: int):
     print(f"Data update product: {req}")
+
+    if not repo.exist_product_id(db=db, id=req.id):
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Product not found",
+        )
+
     return repo.update_product(db=db, req=req, user_id=user_id)
 
 
 def soft_delete_product(req: ProductDelete, db: Session, user_id: int):
     print(f"Delete product id: {req.id}")
+
+    if not repo.exist_product_id(db=db, id=req.id):
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Product not found",
+        )
+
     return repo.delete_product(db=db, id=req.id, user_id=user_id)
